layer: switch to using a watch channel internally
instead of broadcast, watch allows us not to have races between *when*
the message was sent and when a subscriber was created, instead we can
always observe the current state.

use the current state to:
- check if we can evict_and_wait
- complete the evict_and_wait
- notify of when we are non-cancellably "downloading"
- cancel awaiting for eviction if we are racing a "downloading"

eviction path now has an async and a blocking phase in addition to just
a blocking phase, which complicates tests.